(PC-9223) Move idcheck logic into its own feature directory
