+++ src/features/eighteenBirthday/pages/components/EighteenBirthdayCard.tsx
@@ -3,8 +3,8 @@ import { useNavigation } from '@react-navigation/native'
 import React from 'react'
 
 import { useDepositAmount } from 'features/auth/api'
-import { useNavigateToIdCheck } from 'features/auth/signup/idCheck/useNavigateToIdCheck'
 import { useUserProfileInfo } from 'features/home/api'
+import { useNavigateToIdCheck } from 'features/idcheck/hooks/useNavigateToIdCheck'
 import { UseNavigationType } from 'features/navigation/RootNavigator'
 import TutorialPassLogo from 'ui/animations/eighteen_birthday.json'
 import { AchievementCardKeyProps, GenericAchievementCard } from 'ui/components/achievements'
