+++ src/features/idcheck/hooks/useKeyboardAdjustFixIdCheck.d.ts
