+++ src/features/navigation/RootNavigator/routes.ts
@@ -16,8 +16,6 @@ import { AccountCreated } from 'features/auth/signup/AccountCreated'
 import { AfterSignupEmailValidationBuffer } from 'features/auth/signup/AfterSignupEmailValidationBuffer'
 import { BeneficiaryRequestSent } from 'features/auth/signup/BeneficiaryRequestSent'
 import { EligibilityConfirmed } from 'features/auth/signup/EligibilityConfirmed'
-import { IdCheck } from 'features/auth/signup/IdCheck'
-import { IdCheckV2 } from 'features/auth/signup/IdCheckV2'
 import { PhoneValidation } from 'features/auth/signup/PhoneValidation'
 import { PhoneValidationTooManyAttempts } from 'features/auth/signup/PhoneValidationTooManyAttempts'
 import { SetBirthday } from 'features/auth/signup/SetBirthday'
@@ -43,6 +41,8 @@ import { FavoritesSorts } from 'features/favorites/pages/FavoritesSorts'
 import { CulturalSurvey } from 'features/firstLogin/CulturalSurvey'
 import { FirstTutorial } from 'features/firstTutorial/pages/FirstTutorial/FirstTutorial'
 import { ForceUpdate } from 'features/forceUpdate/ForceUpdate'
+import { IdCheck } from 'features/idcheck/IdCheck'
+import { IdCheckV2 } from 'features/idcheck/IdCheckV2'
 import { Maintenance } from 'features/maintenance/Maintenance'
 import { TabNavigator } from 'features/navigation/TabBar/TabNavigator'
 import { Offer, OfferDescription } from 'features/offer'
